refactor(tool): Move config related structs, enums into tool::config module
