+++ iai-callgrind-runner/src/runner/bin_bench.rs
@@ -14,7 +14,8 @@ use super::common::{Assistant, AssistantKind, Baselines, BenchmarkSummaries, Con
 use super::format::{BinaryBenchmarkHeader, OutputFormat};
 use super::meta::Metadata;
 use super::summary::{BaselineKind, BaselineName, BenchmarkKind, BenchmarkSummary, SummaryOutput};
-use super::tool::{RunOptions, ToolConfigs, ToolOutputPath, ToolOutputPathKind};
+use super::tool::config::ToolConfigs;
+use super::tool::{RunOptions, ToolOutputPath, ToolOutputPathKind};
 use crate::api::{
     self, BinaryBenchmarkBench, BinaryBenchmarkConfig, BinaryBenchmarkGroups, DelayKind,
     EntryPoint, Stdin, ValgrindTool,
