+++ iai-callgrind-runner/src/runner/lib_bench.rs
@@ -11,7 +11,8 @@ use super::common::{Assistant, AssistantKind, Baselines, BenchmarkSummaries, Con
 use super::format::{LibraryBenchmarkHeader, OutputFormat};
 use super::meta::Metadata;
 use super::summary::{BaselineKind, BaselineName, BenchmarkKind, BenchmarkSummary, SummaryOutput};
-use super::tool::{RunOptions, ToolConfigs, ToolOutputPath, ToolOutputPathKind};
+use super::tool::config::ToolConfigs;
+use super::tool::{RunOptions, ToolOutputPath, ToolOutputPathKind};
 use crate::api::{
     EntryPoint, LibraryBenchmarkBench, LibraryBenchmarkConfig, LibraryBenchmarkGroups, RawArgs,
     ValgrindTool,
